+++ packages/solidity/src/lib.rs
@@ -2,7 +2,7 @@
 
 #![deny(clippy::nursery, clippy::pedantic, warnings)]
 
-pub mod full;
-//pub mod ics26;
-//pub mod sp1_ics07;
+pub mod ics26;
+pub mod msgs;
+pub mod sp1_ics07;
 //pub mod sp1_ics07_msgs;
